[#12743] VerilatorPlugin: Reimplement the communication layer

+++ src/Plugins/VerilatorPlugin/Verilated/Peripherals/VerilatedUART.cs
@@ -17,7 +17,7 @@ namespace Antmicro.Renode.Peripherals.Verilated
     public class VerilatedUART : BaseDoubleWordVerilatedPeripheral, IUART
     {
         public VerilatedUART(Machine machine, long frequency, string simulationFilePathLinux = null, string simulationFilePathWindows = null, string simulationFilePathMacOS = null,
-            ulong limitBuffer = LimitBuffer, double timeout = DefaultTimeout) : base (machine, frequency, simulationFilePathLinux, simulationFilePathWindows, simulationFilePathMacOS, limitBuffer, timeout)
+            ulong limitBuffer = LimitBuffer, int timeout = DefaultTimeout) : base (machine, frequency, simulationFilePathLinux, simulationFilePathWindows, simulationFilePathMacOS, limitBuffer, timeout)
         {
             IRQ = new GPIO();
         }
