+++ src/Plugins/VerilatorPlugin/VerilatorIntegrationLibrary/src/buses/bus.h
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2010-2019 Antmicro
+// Copyright (c) 2010-2021 Antmicro
 //
 // This file is licensed under the MIT License.
 // Full license text is available in 'licenses/MIT.txt'.
@@ -10,9 +10,9 @@
 struct BaseBus
 {
     public:
-    virtual void tick(bool countEnable, unsigned long steps) = 0;
-    virtual void write(unsigned long addr, unsigned long value) = 0;
-    virtual unsigned long read(unsigned long addr) = 0;
+    virtual void tick(bool countEnable, unsigned long long steps) = 0;
+    virtual void write(unsigned long long addr, unsigned long long value) = 0;
+    virtual unsigned long read(unsigned long long addr) = 0;
     virtual void reset() = 0;
     void (*evaluateModel)();
     unsigned long tickCounter;
